insert diagram

+++ app/src/main/java/by/itacademy/familywallet/model/PieModel.kt
@@ -1,6 +1,6 @@
 package by.itacademy.familywallet.model
 
 class PieModel(
-    val value: Double,
+    val value: Float,
     val category: String?
 )
